navigation + separate methods file
