+++ front/INDEX.html
@@ -14,10 +14,10 @@
               </a>
         
               <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-                <li><a href="home" class="nav-link px-2 link-secondary">Home</a></li>
-                <li><a href="graph" class="nav-link px-2 link-dark">Graph</a></li>
-                <li><a href="theory" class="nav-link px-2 link-dark">Theory</a></li>
-                <li><a href="about" class="nav-link px-2 link-dark">About</a></li>
+                <li><a href="home/" class="nav-link px-2 link-secondary">Home</a></li>
+                <li><a href="graph/" class="nav-link px-2 link-dark">Graph</a></li>
+                <li><a href="theory/" class="nav-link px-2 link-dark">Theory</a></li>
+                <li><a href="about/" class="nav-link px-2 link-dark">About</a></li>
               </ul>
         
               <div class="col-md-3 text-end">
