+++ requirements.txt
@@ -3,3 +3,4 @@ psycopg2>=2.8
 sympy == 1.11
 django-cors-headers >= 3.13
 numpy ~=1.23
+antlr4-python3-runtime ~=4.11
